perf(nav): move navigation to seperate directory
