+++ frontend/App.js
@@ -2,25 +2,7 @@ import React, { useState } from 'react';
 import { Container } from 'native-base';
 import * as Font from 'expo-font';
 import { AppLoading } from 'expo';
-import { createAppContainer } from 'react-navigation';
-import { createStackNavigator } from 'react-navigation-stack';
-import HomeScreen from './src/screens/HomeScreen';
-import AccountsScreen from './src/screens/AccountsScreen';
-import CreateNewAccountScreen from './src/screens/CreateNewAccountScreen';
-
-const RootStack = createStackNavigator(
-  {
-    Home: HomeScreen,
-    Accounts: AccountsScreen,
-    NewAccount: CreateNewAccountScreen,
-  },
-  {
-    initialRouteName: 'Home',
-    headerMode: 'none',
-  },
-);
-
-const AppContainer = createAppContainer(RootStack);
+import AppContainer from './src/navigations/navigation';
 
 export default function App() {
   const [isReady, setIsReady] = useState(false);
